Removes unnecessary fields and logic.
NEXT: Figure out why the fname field is not getting passed to the
calling program. The flags (set in _open) get passed, the return
code from ctx gets passed, but the fname field does not.

I am not sure if the fname field is not getting stored in the map by the
`open` hooks, or if it is getting stored, but not getting copied by the
bpf_user_str() function.

Try adding some kernel printing.